requirements.txt
altgraph==0.17.4
click==8.1.8
macholib==1.16.3
packaging==24.2
pillow==11.1.0
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
pypdf==5.3.1
pypdfium2==4.30.1
setuptools==75.8.2
